requirements.txt
discord.py
yt-dlp
spotipy
PyNaCl
python-dotenv
pytest
pytest-asyncio
